MeshWorker::DoFInfo: consolidate things into a single header.
There are only two functions in the .templates.h file: let's just move
everything over.
