+++ include/deal.II/meshworker/dof_info.templates.h
@@ -16,47 +16,12 @@
 
 #include <deal.II/base/config.h>
 
-#include <deal.II/base/quadrature_lib.h>
-
 #include <deal.II/meshworker/dof_info.h>
 
+DEAL_II_WARNING("This file is deprecated."
+                "Use deal.II/meshworker/dof_info.h instead.")
+
 DEAL_II_NAMESPACE_OPEN
-
-
-namespace MeshWorker
-{
-  template <int dim, int spacedim, typename number>
-  DoFInfo<dim, spacedim, number>::DoFInfo(const BlockInfo &info)
-    : face_number(numbers::invalid_unsigned_int)
-    , sub_number(numbers::invalid_unsigned_int)
-    , block_info(&info, typeid(*this).name())
-    , level_cell(false)
-  {
-    indices_by_block.resize(info.local().size());
-    for (unsigned int i = 0; i < indices_by_block.size(); ++i)
-      indices_by_block[i].resize(info.local().block_size(i));
-  }
-
-
-
-  template <int dim, int spacedim, typename number>
-  void
-  DoFInfo<dim, spacedim, number>::set_block_indices()
-  {
-    for (unsigned int i = 0; i < indices.size(); ++i)
-      {
-        const std::pair<unsigned int, unsigned int> bi =
-          block_info->local().global_to_local(this->block_info->renumber(i));
-        indices_by_block[bi.first][bi.second] = indices_org[i];
-      }
-    // Remove this after
-    // changing block codes
-    for (unsigned int i = 0; i < indices.size(); ++i)
-      indices[this->block_info->renumber(i)] = indices_org[i];
-  }
-} // namespace MeshWorker
-
-
 DEAL_II_NAMESPACE_CLOSE
 
 #endif
