+++ source/meshworker/mesh_worker_info.cc
@@ -16,7 +16,6 @@
 #include <deal.II/lac/trilinos_vector.h>
 #include <deal.II/lac/vector.h>
 
-#include <deal.II/meshworker/dof_info.templates.h>
 #include <deal.II/meshworker/integration_info.templates.h>
 
 DEAL_II_NAMESPACE_OPEN
